Fix query of modified components in a tuple
A query like `(Modified<&A>, Modified<&B>)` should not be skipped if
`A` or `B` is modified. Right now edict does not skip only if all
components is modified.

An empty tuple should not affect any query:
`((Modified<&A>, Modified<&B>), ())` should be skipped if
`A` and `B` is not modified. Right now `()` is always modified and
this query would not be skipped if we fix the previous issue.

Fixes #10